simplify LeafSearchScript
